[ECO-9] Convert all FHERC20 balances to euint64

+++ contracts/test/FHERC20_Harness.sol
@@ -8,11 +8,11 @@ import { FHERC20 } from "../FHERC20.sol";
 contract FHERC20_Harness is FHERC20 {
     constructor(string memory name_, string memory symbol_, uint8 decimals_) FHERC20(name_, symbol_, decimals_) {}
 
-    function mint(address account, uint128 value) public {
+    function mint(address account, uint64 value) public {
         _mint(account, value);
     }
 
-    function burn(address account, uint128 value) public {
+    function burn(address account, uint64 value) public {
         _burn(account, value);
     }
